feat: topple the monolith; lint its remains
