+++ .node-version
@@ -1,0 +1,1 @@
+22.15.0
