+++ src/index.mjs
@@ -1,0 +1,4 @@
+import { ForgeVTT } from "./ForgeVTT.mjs";
+import "./hooks.mjs";
+
+ForgeVTT.setupForge();
